+++ klum-ast/src/main/java/com/blackbuild/groovy/configdsl/transform/ast/DslAstHelper.java
@@ -140,13 +140,13 @@ public class DslAstHelper {
         return storeAndReturn(fieldNode, ELEMENT_NAME_METADATA_KEY, LanguageHelper.getSingularForPlural(fieldNode.getName()));
     }
 
-    static boolean isKeyed(ClassNode type) {
+    public static boolean isKeyed(ClassNode type) {
         if (!isDSLObject(type))
             return false;
         return getKeyField(type) != null;
     }
 
-    static FieldNode getKeyField(ClassNode target) {
+    public static FieldNode getKeyField(ClassNode target) {
         FieldNode result = target.getNodeMetaData(KEY_FIELD_METADATA_KEY);
 
         if (result == NO_SUCH_FIELD)